Refactor: pagination 유틸 분리#3
